requirements.txt
Django==2.1.15
django-crispy-forms==1.11.1
djangorestframework==3.9.4
gunicorn==20.0.4
Pillow==9.0.0
psycopg2==2.8.6
pytz==2021.1
